Rescheduling in wca-scheduler. Score algorithm.

+++ wca/scheduler/simulator_experiments/generic_experiment.py
@@ -16,9 +16,9 @@ import itertools
 import logging
 import os
 import shutil
-from typing import List, Tuple, Callable, Dict, Type
+from typing import List, Tuple, Callable, Dict, Type, Union
 
-from wca.nodes import Node, Task
+from wca.nodes import Node
 from wca.scheduler.algorithms import Algorithm
 from wca.scheduler.cluster_simulator import ClusterSimulator
 from wca.scheduler.data_providers.cluster_simulator_data_provider import \
@@ -33,46 +33,48 @@ log = logging.getLogger(__name__)
 reports_root_directory: str = 'experiments_results'
 
 
-def experiments_iterator(exp_name,
-                         lengths: List[int], # max iterations
+def experiments_iterator(exp_name, simulator_args,
+                         lengths: List[int],  # max iterations
                          task_gen_func_defs: List[Tuple[Type[TaskGenerator], dict]],
                          nodes_sets: List[List[Node]],
-                         algorithm_defs: List[Tuple[Type[Algorithm], dict]]
+                         algorithm_defs: List[Tuple[Type[Algorithm], dict]],
+                         rmtree: bool = True,
+                         charts: bool = False,
+                         metrics: Union[bool, List[str]] = False,
                          ):
-
     exp_stats: List[dict] = []
 
     # Recreate results directory.
     exp_dir = '{}/{}'.format(reports_root_directory, exp_name)
-    if os.path.isdir(exp_dir):
-        shutil.rmtree(exp_dir)
-    os.makedirs(exp_dir)
+    if rmtree:
+        if os.path.isdir(exp_dir):
+            shutil.rmtree(exp_dir)
+    os.makedirs(exp_dir, exist_ok=True)
 
-    for exp_iter, args in enumerate(itertools.product(
-        lengths,
-        task_gen_func_defs,
-        nodes_sets,
-        algorithm_defs
+    for n, args in enumerate(itertools.product(
+            lengths,
+            task_gen_func_defs,
+            nodes_sets,
+            algorithm_defs
     )):
         length, task_gen_func_def, nodes, algorithm_def = args
         iterations_data, task_gen, simulator = \
-            perform_one_experiment(length, task_gen_func_def, nodes, algorithm_def)
+            perform_one_experiment(simulator_args, length, task_gen_func_def, nodes, algorithm_def)
 
         # Sub experiment report
-        subexp_title = '%d_%snodes_%s_%s' % (
-        exp_iter, len(simulator.nodes), task_gen, simulator.algorithm)
+        subexp_name = '%d_%snodes_%s_%s' % (n, len(simulator.nodes), task_gen, simulator.algorithm)
         stats = generate_subexperiment_report(
             exp_name,
             exp_dir,
-            subexp_title,
+            subexp_name,
             iterations_data,
             task_gen=task_gen,
             algorithm=simulator.algorithm,
-            charts=False,
-            extra_charts=False,
+            charts=charts,
+            metrics=metrics,
         )
 
-        log.debug('Finished experiment.', exp_iter)
+        log.debug('Finished experiment.', n)
         log.debug('Stats:', stats)
         exp_stats.append(stats)
 
@@ -81,8 +83,9 @@ def experiments_iterator(exp_name,
 
 
 def perform_one_experiment(
+        simulator_args: dict,
         length: int,
-        task_generator_def: Tuple[Callable, Dict],
+        task_gen_def: Tuple[Type[TaskGenerator], Dict],
         nodes: List[Node],
         algorithm_def: Tuple[Type[Algorithm], Dict],
 ):
@@ -90,11 +93,11 @@ def perform_one_experiment(
     iterations_data: List[IterationData] = []
 
     # Back reference between data proxy and simulator.
-    simulator = ClusterSimulator(tasks=[], nodes=nodes, algorithm=None)
+    simulator = ClusterSimulator(tasks=[], nodes=nodes, algorithm=None, **simulator_args)
     data_proxy = ClusterSimulatorDataProvider(simulator)
     simulator.algorithm = algorithm_class(data_provider=data_proxy, **algorithm_args)
 
-    task_gen_class, task_gen_args = task_generator_def
+    task_gen_class, task_gen_args = task_gen_def
     task_gen = task_gen_class(**task_gen_args)
 
     run_n_iter(length, simulator, task_gen,
@@ -105,11 +108,13 @@ def perform_one_experiment(
 
 def run_n_iter(length: int,
                simulator: ClusterSimulator,
-               task_gen: Callable[[int], Task],
+               task_gen: TaskGenerator,
                iteration_finished_callback: Callable):
     iteration = 0
     while iteration < length:
-        simulator.iterate_single_task(task_gen(iteration))
+        new_task = task_gen(iteration)
+        assert new_task is not None or iteration > 0, 'in first iteration we need at least one task'
+        simulator.iterate_single_task(new_task)
         iteration_finished_callback(iteration, simulator)
         iteration += 1
     return simulator
